requirements.txt
asgiref==3.6.0
certifi==2022.12.7
distlib==0.3.6
Django==4.2
django-crontab==0.7.1
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==18.4.0
filelock==3.11.0
pipenv==2023.3.20
platformdirs==3.2.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
virtualenv==20.21.0
virtualenv-clone==0.5.7
